+++ src/lib/structs/Slash.sol
@@ -2,6 +2,6 @@
 pragma solidity 0.8.22;
 
 struct Slash {
-  uint16 reservePoolId;
+  uint8 reservePoolId;
   uint256 amount;
 }
